Move waitExpect to test
