+++ universal-login-commons/test/core/utils/wait.ts
@@ -1,6 +1,7 @@
 import chai, {expect} from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import {waitUntil, waitExpect, sleep} from '../../../lib/core/utils/wait';
+import {waitUntil, sleep} from '../../../lib/core/utils/wait';
+import {waitExpect} from '../../helpers/waitExpect';
 
 chai.use(chaiAsPromised);
