+++ universal-login-commons/test/index.ts
@@ -2,3 +2,4 @@ import MockToken from './fixtures/MockToken.json';
 export {MockToken};
 export {deployMockToken} from './fixtures/deployMockToken';
 export {basicENS, deployENS} from './fixtures/basicENS';
+export {waitExpect} from './helpers/waitExpect';
